adaptative text almost work

+++ include/battle_interface.h
@@ -108,7 +108,7 @@ void TryToAddMoveInfoWindow(void);
 void TryToAddEnemyInfoWindow(void);
 void TryToHideEnemyInfoWindow(void);
 
-void CreateExtraSkillPopUp(u32 extraskill);
+void CreateExtraSkillPopUp(u8 extraskill);
 void DestroyExtraSkillPopUp();
 
 #endif // GUARD_BATTLE_INTERFACE_H
